Window Functions

Several kinds of window functions

IndexFunArrays.window_radial_linear — Function

window_radial_linear([T=Float64], size::NTuple; 
            offset=CtrFT, scale=ScaFTEdge, border_in=0.8, border_out=1.0)

A multidimensional radial window with a linear transition from zero at the borders (border_out) to one (border_in). With the default offset and scale the borders are specified relative to the edge.

julia> window_radial_linear((4,5),border_in=0.0)
4×5 IndexFunArray{Float64, 2, IndexFunArrays.var"#59#60"{Float64, Tuple{Float64, Float64}, Tuple{Float64, Float64}, Float64, Float64}}:
 0.0  0.0       0.0  0.0       0.0
 0.0  0.292893  0.5  0.292893  0.0
 0.0  0.5       1.0  0.5       0.0
 0.0  0.292893  0.5  0.292893  0.0

IndexFunArrays.window_hanning — Function

window_hanning([T=Float64], size::NTuple; 
            offset=CtrFT, scale=ScaFTEdge, border_in=0.8, border_out=1.0)

A multidimensional (separable) window with a von Hann transition between the borders (border_out) to one (border_in). See ?window_linear for more details on the arguments.

IndexFunArrays.window_radial_hamming — Function

window_radial_hamming([T=Float64], size::NTuple; 
            offset=CtrFT, scale=ScaFTEdge, border_in=0.8, border_out=1.0)

A multidimensional radial window with a Hamming transition between the borders (border_out) to one (border_in). See ?window_radial_linear for more details on the arguments.

IndexFunArrays.window_hamming — Function

window_hamming([T=Float64], size::NTuple; 
            offset=CtrFT, scale=ScaFTEdge, border_in=0.8, border_out=1.0)

A multidimensional (separable) window with a Hamming transition between the borders (border_out) to one (border_in). See ?window_linear for more details on the arguments.

IndexFunArrays.window_radial_hanning — Function

window_radial_hanning([T=Float64], size::NTuple; 
            offset=CtrFT, scale=ScaFTEdge, border_in=0.8, border_out=1.0)

A multidimensional radial window with a von Hann transition between the borders (border_out) to one (border_in). See ?window_radial_linear for more details on the arguments.

IndexFunArrays.window_edge — Function

window_edge([T=Float64], size::NTuple; 
            offset=CtrFT, scale=ScaFTEdge, border_in=0.8, border_out=1.0)

A multidimensional (separable) window with a sudden transition half way between the borders (border_out) to one (border_in). See ?window_linear for more details on the arguments.

IndexFunArrays.window_radial_blackman_harris — Function

window_radial_blackman_harris([T=Float64], size::NTuple; 
            offset=CtrFT, scale=ScaFTEdge, border_in=0.8, border_out=1.0)

A multidimensional radial window with a Hamming transition according to Blackman/Harris between the borders (border_out) to one (border_in). See ?window_radial_linear for more details on the arguments.

IndexFunArrays.window_blackman_harris — Function

window_blackman_harris([T=Float64], size::NTuple; 
            offset=CtrFT, scale=ScaFTEdge, border_in=0.8, border_out=1.0)

A multidimensional (separable) window with a transition according to Blackman/Harris between the borders (border_out) to one (border_in). See ?window_linear for more details on the arguments.

IndexFunArrays.window_radial_edge — Function

window_radial_edge([T=Float64], size::NTuple; 
            offset=CtrFT, scale=ScaFTEdge, border_in=0.8, border_out=1.0)

A multidimensional radial window (disk) with a sudden transition half way between the borders (border_out) to one (border_in). See ?window_radial_linear for more details on the arguments.

IndexFunArrays.window_linear — Function

window_linear([T=Float64], size::NTuple; 
            offset=CtrFT, scale=ScaFTEdge, border_in=0.8, border_out=1.0)

A multidimensional (separable) window with a linear transition from zero at the borders (border_out) to one (border_in).

julia> window_linear((8,9),border_in=0.0)
8×9 IndexFunArray{Float64, 2, IndexFunArrays.var"#34#35"{Float64, Tuple{Float64, Float64}, Tuple{Float64, Float64}, Float64, Float64}}:
 0.0  0.0     0.0    0.0     0.0   0.0     0.0    0.0     0.0
 0.0  0.0625  0.125  0.1875  0.25  0.1875  0.125  0.0625  0.0
 0.0  0.125   0.25   0.375   0.5   0.375   0.25   0.125   0.0
 0.0  0.1875  0.375  0.5625  0.75  0.5625  0.375  0.1875  0.0
 0.0  0.25    0.5    0.75    1.0   0.75    0.5    0.25    0.0
 0.0  0.1875  0.375  0.5625  0.75  0.5625  0.375  0.1875  0.0
 0.0  0.125   0.25   0.375   0.5   0.375   0.25   0.125   0.0
 0.0  0.0625  0.125  0.1875  0.25  0.1875  0.125  0.0625  0.0